[IMP] do not delete when still used in extra module

+++ openupgrade_scripts/scripts/mrp/14.0.2.0/pre-migration.py
@@ -119,12 +119,13 @@ def fill_mrp_routing_workcenter_bom_id(env):
         JOIN mrp_routing_workcenter mrw2 ON mrw.old_routing_workcenter_id = mrw2.id
         WHERE sm.raw_material_production_id = mp.id AND sm.operation_id = mrw2.id""",
     )
-    openupgrade.logged_query(
-        env.cr,
-        """
-        DELETE FROM mrp_routing_workcenter
-        WHERE old_routing_workcenter_id IS NULL""",
-    )
+    if not openupgrade.table_exists(env.cr, "mrp_bom_operation"):
+        openupgrade.logged_query(
+            env.cr,
+            """
+            DELETE FROM mrp_routing_workcenter
+            WHERE old_routing_workcenter_id IS NULL""",
+        )
 
 
 @openupgrade.migrate()
